backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
psycopg2-binary==2.9.10
gunicorn==23.0.0
python-dotenv==1.0.1
mcp==1.9.2
# Document processing
python-docx==1.1.2
PyPDF2==3.0.1
pdfplumber==0.11.6
# NLP / Detection
nltk==3.9.1
textstat==0.7.4
scikit-learn==1.6.1
numpy==2.2.3
scipy==1.15.2
# AI providers
anthropic==0.49.0
# Testing
pytest==8.3.4
